+++ src/main/java/eu/gir/girsignals/init/GIRBlocks.java
@@ -13,6 +13,7 @@ import eu.gir.girsignals.blocks.RedstoneInput;
 import eu.gir.girsignals.blocks.Signal;
 import eu.gir.girsignals.blocks.SignalBox;
 import eu.gir.girsignals.blocks.SignalController;
+import eu.gir.girsignals.blocks.boards.SignalAndreasCross;
 import eu.gir.girsignals.blocks.boards.SignalBUE;
 import eu.gir.girsignals.blocks.boards.SignalBUELight;
 import eu.gir.girsignals.blocks.boards.SignalEL;
@@ -27,6 +28,8 @@ import eu.gir.girsignals.blocks.signals.SignalHL;
 import eu.gir.girsignals.blocks.signals.SignalHV;
 import eu.gir.girsignals.blocks.signals.SignalKS;
 import eu.gir.girsignals.blocks.signals.SignalSHLight;
+import eu.gir.girsignals.blocks.signals.SignalSHMech;
+import eu.gir.girsignals.blocks.signals.SignalSemaphore;
 import eu.gir.girsignals.blocks.signals.SignalTram;
 import net.minecraft.block.Block;
 import net.minecraft.block.ITileEntityProvider;
@@ -69,7 +72,10 @@ public final class GIRBlocks {
     public static final SignalBox SIGNAL_BOX = new SignalBox();
     public static final RedstoneIO REDSTONE_IN = new RedstoneInput();
     public static final RedstoneIO REDSTONE_OUT = new RedstoneIO();
-
+    public static final SignalSemaphore SEMAPHORE_SIGNAL = new SignalSemaphore();
+    public static final SignalSHMech SH_MECH = new SignalSHMech();
+    public static final SignalAndreasCross ANDREAS_CROSS = new SignalAndreasCross();
+    
     public static ArrayList<Block> blocksToRegister = new ArrayList<>();
 
     public static void init() {