+++ src/main/java/eu/gir/girsignals/signalbox/config/RSSignalConfig.java
@@ -14,6 +14,7 @@ import eu.gir.girsignals.blocks.signals.SignalHL;
 import eu.gir.girsignals.blocks.signals.SignalHV;
 import eu.gir.girsignals.blocks.signals.SignalKS;
 import eu.gir.girsignals.blocks.signals.SignalSHLight;
+import eu.gir.girsignals.blocks.signals.SignalSemaphore;
 import eu.gir.girsignals.enums.PathType;
 import eu.gir.girsignals.tileentitys.SignalTileEnity;
 
@@ -36,6 +37,7 @@ public final class RSSignalConfig implements ISignalAutoconfig {
         values.put(SignalKS.MAINSIGNAL, KSMain.KS_SHUNTING);
         values.put(SignalHL.STOPSIGNAL, HL.HL_SHUNTING);
         values.put(SignalHL.EXITSIGNAL, HLExit.HL_SHUNTING);
+        values.put(SignalSemaphore.RA12, true);
 
         if (info.type.equals(PathType.SHUNTING)) {
             values.put(SignalRA.RALIGHT, true);